fix: устранить ошибки

+++ src/main/java/ru/practicum/shareit/user/UserController.java
@@ -2,15 +2,20 @@ package ru.practicum.shareit.user;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.shareit.user.dto.UserDto;
 import ru.practicum.shareit.user.service.UserService;
 
+import javax.validation.ConstraintViolationException;
 import javax.validation.Valid;
+import javax.validation.constraints.Min;
+import java.util.HashSet;
 import java.util.List;
 
 
 @Slf4j
+@Validated
 @RestController
 @RequiredArgsConstructor
 @RequestMapping(path = "/users")
@@ -27,7 +32,12 @@ public class UserController {
     }
 
     @PatchMapping("/{userId}")
-    public UserDto updateUser(@PathVariable Integer userId, @Valid @RequestBody UserDto userDto) {
+    public UserDto updateUser(@PathVariable @Min(1) Integer userId, @Valid @RequestBody UserDto userDto)
+            throws ConstraintViolationException {
+        if (userId < 1) {
+            throw new ConstraintViolationException("Invalid userId", new HashSet<>());
+        }
+
         User user = userMapper.toUser(userDto);
         User updatedUser = service.update(userId, user);
         log.info("Update user {}.", updatedUser);
@@ -41,14 +51,22 @@ public class UserController {
     }
 
     @GetMapping("/{userId}")
-    public UserDto getUserById(@PathVariable Integer userId) {
+    public UserDto getUserById(@PathVariable @Min(1) Integer userId) throws ConstraintViolationException {
+        if (userId < 1) {
+            throw new ConstraintViolationException("Invalid userId", new HashSet<>());
+        }
+
         User user = service.getData(userId);
         log.info("Get user with id = {}.", userId);
         return userMapper.toUserDto(user);
     }
 
     @DeleteMapping("/{userId}")
-    public void deleteUserById(@PathVariable Integer userId) {
+    public void deleteUserById(@PathVariable @Min(1) Integer userId) throws ConstraintViolationException {
+        if (userId < 1) {
+            throw new ConstraintViolationException("Invalid userId", new HashSet<>());
+        }
+
         service.delete(userId);
         log.info("User with id = {} was deleted.", userId);
     }
